requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Data handling
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging and tracking
tensorboard>=2.13.0
tqdm>=4.65.0

# Metaflow for inference pipeline (bonus)
metaflow>=2.10.0

# Configuration management
pyyaml>=6.0.0

# Optional: For DOTA dataset parsing
xmltodict>=0.13.0

